requirements.txt
async-timeout==4.0.2
azure-core==1.27.1
azure-storage-blob==12.16.0
blinker==1.6.2
blis==0.7.9
blob==0.16
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clean-text==0.6.0
cleantext==1.1.4
click==8.1.3
colorama==0.4.6
confection==0.1.0
cryptography==41.0.1
cymem==2.0.7
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
Flask==2.3.2
ftfy==6.1.1
idna==3.4
importlib-metadata==6.7.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.0
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==1.5.3
pathy==0.10.2
preshed==3.0.8
pycparser==2.21
pydantic==1.10.9
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
threadpoolctl==3.1.0
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.15.0
